Initial source files for the project
The project has been divided into couple of packages depending upon the
division of responsibilities of the modules.

MODEL:
This should contain the objects that have to be worked upon. In our
project, this should be the images, videos etc.

VIEW:
The classes in this package are responsible for presenting the GUI to
the user.

CONTROLLER:
The classes here would control the execution of the application. They
would be solely responsible for driving the logic of computations on the
MODEL objects (images and videos), handling the GUI events and using the
VIEW to control the display outputs