Video file handling enabled partially
Added some more code (small part) for handling the video files.

With this code you should be able to open the video file in the right
hand side of the window
You should be able to traverse through the frames on the screen by
pressing  ">>" and "<<" buttons, once the video is displayed. The video
is displayed initially with its first frame shown on the screen.
PLAY/PAUSE is not implemented now, but for the image search logic, its
nor needed immediately and can be handled later too.