Some more code added
Added the logic for further recursion, diving down the quadrants again
into 9 quadrants for search.

Positive test cases almost work. For orange, the girl eating half orange
is not get selected, probably because in her classroom env, the mixture
of RGB is similar to that in the orange... Works for rest.

The negative test cases are failing now.... If I reduce the fault
tolerance, then the positive test cases have chances of failing.. :(